+++ examples/cindel_todo/lib/features/todos/data/datasources/todos_local_data_source.dart
@@ -32,7 +32,7 @@ final class TodosLocalDataSource {
     final database = await _database;
     final todos = await database.todos
         .where()
-        .titleStartsWith(prefix)
+        .titleWordsStartsWith(prefix)
         .findAll();
     return _sortModels(todos);
   }